+++ packages/next/src/client/components/work-async-storage-instance.ts
@@ -1,4 +1,4 @@
-import type { StaticGenerationAsyncStorage } from './static-generation-async-storage.external'
+import type { StaticGenerationAsyncStorage } from './work-async-storage.external'
 import { createAsyncLocalStorage } from './async-local-storage'
 
 export const staticGenerationAsyncStorage: StaticGenerationAsyncStorage =
