+++ packages/next/src/client/components/work-async-storage.external.ts
@@ -6,7 +6,7 @@ import type { Revalidate } from '../../server/lib/revalidate'
 import type { FallbackRouteParams } from '../../server/request/fallback-params'
 
 // Share the instance module in the next-shared layer
-import { staticGenerationAsyncStorage } from './static-generation-async-storage-instance' with { 'turbopack-transition': 'next-shared' }
+import { staticGenerationAsyncStorage } from './work-async-storage-instance' with { 'turbopack-transition': 'next-shared' }
 import type { AppSegmentConfig } from '../../build/app-segments/app-segment-config'
 
 export interface StaticGenerationStore {
