+++ packages/next/src/server/async-storage/with-work-store.ts
@@ -1,11 +1,11 @@
 import type { WithStore } from './with-store'
-import type { StaticGenerationStore } from '../../client/components/static-generation-async-storage.external'
+import type { StaticGenerationStore } from '../../client/components/work-async-storage.external'
 import type { AsyncLocalStorage } from 'async_hooks'
 import type { IncrementalCache } from '../lib/incremental-cache'
 import type { RenderOptsPartial } from '../app-render/types'
 import type { FetchMetric } from '../base-http'
 import type { RequestLifecycleOpts } from '../base-server'
-import type { FallbackRouteParams } from '../../server/request/fallback-params'
+import type { FallbackRouteParams } from '../request/fallback-params'
 import type { AppSegmentConfig } from '../../build/app-segments/app-segment-config'
 
 import { normalizeAppPath } from '../../shared/lib/router/utils/app-paths'
